# src/app/docs/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
...
  controller.CatItem:
    properties:
      adoption_price:
        type: integer
      cat_thumbnail_path:
        type: string
      contract_benefit:
        type: integer
      contract_days:
        type: integer
      deposit:
        type: integer
      id:
        type: integer
      level:
        type: string
      name:
        type: string
      pet_coin:
        type: integer
      price:
        type: integer
      reservation_price:
        type: integer
      status:
        type: integer
    type: object
...
  controller.GetCatListResponse:
    properties:
      cats:
        items:
          $ref: '#/definitions/controller.CatItem'
        type: array
      lower:
        type: integer
      total:
        type: integer
      upper:
        type: integer
    type: object
...
  controller.Message:
    properties:
      message:
        example: message
        type: string
    type: object
...
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
host: '{{.Host}}'
info:
  contact: {}
  description: '{{.Description}}'
  license: {}
  title: '{{.Title}}'
  version: '{{.Version}}'
paths:
...
  /api/v1/adoption/cats:
    get:
      consumes:
      - application/json
      description: get adoption cat list with status from database
      parameters:
      - description: 貓列表的lower
        in: query
        name: lower
        required: true
        type: integer
      - description: 貓列表的upper
        in: query
        name: upper
        required: true
        type: integer
      - description: '貓列表的領養狀態(  增值中: 1 / 已完成 :2 / 已售出 : 3)'
        in: query
        name: status
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetCatListResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
...
  /api/v1/cat:
    post:
      consumes:
      - multipart/form-data
      description: Add a new cat
      parameters:
      - description: 貓的名稱
        in: formData
        name: name
        required: true
        type: string
      - description: 貓的級別
        in: formData
        name: level
        required: true
        type: string
      - description: 貓的價格
        in: formData
        name: price
        required: true
        type: string
      - description: 貓的押金
        in: formData
        name: deposit
        required: true
        type: string
      - description: 貓的pet幣
        in: formData
        name: pet_coin
        required: true
        type: string
      - description: 貓的預約價格
        in: formData
        name: reservation_price
        required: true
        type: string
      - description: 貓的即搶價格
        in: formData
        name: adoption_price
        required: true
        type: string
      - description: 貓的合約時間
        in: formData
        name: contract_days
        required: true
        type: string
      - description: 貓的合約增益
        in: formData
        name: contract_benefit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Message'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
...
  /api/v1/cat/thumbnails/{thumbnailId}/cats/{catId}:
    post:
      consumes:
      - application/json
      description: link cat thumbnail with cat to the database
      parameters:
      - description: 貓的ID
        in: path
        name: catId
        required: true
        type: integer
      - description: 貓的縮圖ID
        in: path
        name: thumbnailId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Message'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
  /api/v1/cats:
    get:
      consumes:
      - application/json
      description: get cat list with status from database
      parameters:
      - description: 貓列表的lower
        in: query
        name: lower
        required: true
        type: integer
      - description: 貓列表的upper
        in: query
        name: upper
        required: true
        type: integer
      - description: '貓列表的狀態(系統掛售中 : 1 / 期約到期掛售中 : 2/轉讓中 : 3 /領養增值中 :4 / 等待裂變中 :5)'
        in: query
        name: status
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetCatListResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
  /api/v1/cats/{catId}:
    put:
      consumes:
      - multipart/form-data
      description: modify cat information to the database
      parameters:
      - description: 貓的ID
        in: path
        name: catId
        required: true
        type: integer
      - description: 貓的名稱
        in: formData
        name: name
        required: true
        type: string
      - description: 貓的級別
        in: formData
        name: level
        required: true
        type: string
      - description: 貓的價格
        in: formData
        name: price
        required: true
        type: string
      - description: 貓的押金
        in: formData
        name: deposit
        required: true
        type: string
      - description: 貓的pet幣
        in: formData
        name: pet_coin
        required: true
        type: string
      - description: 貓的預約價格
        in: formData
        name: reservation_price
        required: true
        type: string
      - description: 貓的即搶價格
        in: formData
        name: adoption_price
        required: true
        type: string
      - description: 貓的合約時間
        in: formData
        name: contract_days
        required: true
        type: string
      - description: 貓的合約增益
        in: formData
        name: contract_benefit
        required: true
        type: string
      - description: '貓的狀態(系統掛售中 : 1 / 期約到期掛售中 : 2/轉讓中 : 3 /領養增值中 :4 / 等待裂變中 :5)'
        in: formData
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Message'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
...
  /api/v1/time/schedules/{scheduleId}/cat/{catId}:
    post:
      consumes:
      - application/json
      description: Add a new timePeriodCatPivot
      parameters:
      - description: 貓的ID
        in: path
        name: catId
        required: true
        type: integer
      - description: 時段的ID
        in: path
        name: scheduleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Message'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
...
  /api/v1/transfer/cats:
    get:
      consumes:
      - application/json
      description: get transfer cat list with status from database
      parameters:
      - description: 貓列表的lower
        in: query
        name: lower
        required: true
        type: integer
      - description: 貓列表的upper
        in: query
        name: upper
        required: true
        type: integer
      - description: '貓列表的交易狀態( 待交易: 1 / 買家已上傳憑證 : 2/已完成 :3 /已取消)'
        in: query
        name: status
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetCatListResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
...
swagger: "2.0"